+++ package.json
@@ -1,0 +1,23 @@
+{
+  "name": "hanggliding",
+  "version": "1.0.0",
+  "description": "",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/KirilKaloyanov/hanggliding.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/KirilKaloyanov/hanggliding/issues"
+  },
+  "homepage": "https://github.com/KirilKaloyanov/hanggliding#readme",
+  "dependencies": {
+    "express": "^4.18.2"
+  }
+}
